+++ GlkView/GlkCustomTextSection.h
@@ -22,7 +22,7 @@ extern NSString* GlkCustomSectionAttributeName;
 
 // Note that this element is not typeset when this is called (wait for the placeBaselineAt: call)
 /// Request from the typesetter that this object generate a suitable line section object. Returns YES to indicate that a line section has been created
-- (BOOL) formatSectionAtOffset: (float) offset
+- (BOOL) formatSectionAtOffset: (CGFloat) offset
 				  inTypesetter: (GlkTypesetter*) typesetter
 				 forGlyphRange: (NSRange) glyphs;
